Released version works, comments added, and balance
Modification of Enemy::slow(**) and Enemy::unslow()

+++ TowerDefense/Config.h
@@ -134,7 +134,7 @@ extern const string		SLOW_TOWER_SPRITE_ADD;
 /*
 Attack constants
 */
-extern const int		SLOW_EFFECT;
+extern const float		SLOW_EFFECT;
 
 
 /*
